requirements.txt
requests 
beautifulsoup4 
pandas
lxml
selenium
webdriver-manager
flask
matplotlib 
seaborn 
openai
fpdf2
